full unit test coverage on model_wrapper, minor bug fixes
